requirements.txt
 docx
 pillow
 jazzmin
 exception
 whitenoise
 fitz
 daphne
 channels
 tools
 mysqlclient
 filetype
 agora_token_builder
 groq
 reportlab
 